UX-644 convert ProgressBar to typescript
